requirements.txt
biopython==1.76 ; python_version <= '2.7' or (python_version >= '3.0' and python_version <= '3.5')
biopython>=1.77 ; (python_version > '2.7' and python_version < '3.0') or python_version > '3.5'
future>=0.18.2
matplotlib>=2.2.5
networkx>=2.2
numpy>=1.16.6
Pillow>=6.2.2
python-dateutil>=2.8.1
six>=1.14.0
